Add possibility to stream input geometry in smaller chunks
Change the way recast is fed with polygon soup so that the geometry can
be streamed.
