+++ detour/src/test/java/org/recast4j/detour/RecastTestMeshBuilder.java
@@ -17,7 +17,6 @@ freely, subject to the following restrictions:
 */
 package org.recast4j.detour;
 
-import org.recast4j.recast.InputGeom;
 import org.recast4j.recast.ObjImporter;
 import org.recast4j.recast.PolyMesh;
 import org.recast4j.recast.PolyMeshDetail;
@@ -26,6 +25,7 @@ import org.recast4j.recast.RecastBuilder.RecastBuilderResult;
 import org.recast4j.recast.RecastBuilderConfig;
 import org.recast4j.recast.RecastConfig;
 import org.recast4j.recast.RecastConstants.PartitionType;
+import org.recast4j.recast.geom.InputGeomProvider;
 
 public class RecastTestMeshBuilder {
 
@@ -51,7 +51,7 @@ public class RecastTestMeshBuilder {
 				m_detailSampleMaxError);
 	}
 
-	public RecastTestMeshBuilder(InputGeom m_geom, PartitionType m_partitionType, float m_cellSize, float m_cellHeight,
+	public RecastTestMeshBuilder(InputGeomProvider m_geom, PartitionType m_partitionType, float m_cellSize, float m_cellHeight,
 			float m_agentHeight, float m_agentRadius, float m_agentMaxClimb, float m_agentMaxSlope, int m_regionMinSize,
 			int m_regionMergeSize, float m_edgeMaxLen, float m_edgeMaxError, int m_vertsPerPoly,
 			float m_detailSampleDist, float m_detailSampleMaxError) {
