+++ detour/src/test/java/org/recast4j/detour/io/MeshSetReaderWriterTest.java
@@ -15,13 +15,13 @@ import org.recast4j.detour.MeshTile;
 import org.recast4j.detour.NavMesh;
 import org.recast4j.detour.SampleAreaModifications;
 import org.recast4j.detour.TestDetourBuilder;
-import org.recast4j.recast.InputGeom;
 import org.recast4j.recast.ObjImporter;
 import org.recast4j.recast.Recast;
 import org.recast4j.recast.RecastBuilder;
 import org.recast4j.recast.RecastBuilderConfig;
 import org.recast4j.recast.RecastConfig;
 import org.recast4j.recast.RecastConstants.PartitionType;
+import org.recast4j.recast.geom.InputGeomProvider;
 
 public class MeshSetReaderWriterTest {
 
@@ -47,7 +47,7 @@ public class MeshSetReaderWriterTest {
 	@Test
 	public void test() throws IOException {
 
-		InputGeom geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
+		InputGeomProvider geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
 
 		NavMeshSetHeader header = new NavMeshSetHeader();
 		header.magic = NavMeshSetHeader.NAVMESHSET_MAGIC;
