+++ detourtilecache/src/test/java/org/recast4j/detour/tilecache/TempObstaclesTest.java
@@ -9,16 +9,16 @@ import java.util.List;
 
 import org.junit.Test;
 import org.recast4j.detour.MeshTile;
-import org.recast4j.recast.InputGeom;
 import org.recast4j.recast.ObjImporter;
 import org.recast4j.recast.RecastBuilder;
+import org.recast4j.recast.geom.InputGeomProvider;
 
 public class TempObstaclesTest extends AbstractTileCacheTest {
 
 	@Test
 	public void testDungeon() throws IOException {
 		boolean cCompatibility = true;
-		InputGeom geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
+		InputGeomProvider geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
 		TestTileLayerBuilder layerBuilder = new TestTileLayerBuilder(geom);
 		List<byte[]> layers = layerBuilder.build(ByteOrder.LITTLE_ENDIAN, cCompatibility, 1);
 		TileCache tc = getTileCache(geom, ByteOrder.LITTLE_ENDIAN, cCompatibility);
@@ -49,7 +49,7 @@ public class TempObstaclesTest extends AbstractTileCacheTest {
 	@Test
 	public void testDungeonBox() throws IOException {
 		boolean cCompatibility = true;
-		InputGeom geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
+		InputGeomProvider geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
 		TestTileLayerBuilder layerBuilder = new TestTileLayerBuilder(geom);
 		List<byte[]> layers = layerBuilder.build(ByteOrder.LITTLE_ENDIAN, cCompatibility, 1);
 		TileCache tc = getTileCache(geom, ByteOrder.LITTLE_ENDIAN, cCompatibility);
