+++ detourtilecache/src/test/java/org/recast4j/detour/tilecache/io/TileCacheReaderWriterTest.java
@@ -15,9 +15,9 @@ import org.recast4j.detour.MeshTile;
 import org.recast4j.detour.tilecache.AbstractTileCacheTest;
 import org.recast4j.detour.tilecache.TestTileLayerBuilder;
 import org.recast4j.detour.tilecache.TileCache;
-import org.recast4j.recast.InputGeom;
 import org.recast4j.recast.ObjImporter;
 import org.recast4j.recast.RecastBuilder;
+import org.recast4j.recast.geom.InputGeomProvider;
 
 public class TileCacheReaderWriterTest extends AbstractTileCacheTest {
 
@@ -37,7 +37,7 @@ public class TileCacheReaderWriterTest extends AbstractTileCacheTest {
 	}
 
 	private void testDungeon(boolean cCompatibility) throws IOException {
-		InputGeom geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
+		InputGeomProvider geom = new ObjImporter().load(RecastBuilder.class.getResourceAsStream("dungeon.obj"));
 		TestTileLayerBuilder layerBuilder = new TestTileLayerBuilder(geom);
 		List<byte[]> layers = layerBuilder.build(ByteOrder.LITTLE_ENDIAN, cCompatibility, 1);
 		TileCache tc = getTileCache(geom, ByteOrder.LITTLE_ENDIAN, cCompatibility);
